requirements.txt
Flask
weaviate-client
ollama
langchain